web_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
requests==2.31.0
chromadb>=0.4.18
pypdf2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
python-magic==0.4.27
typing-extensions==4.8.0
pydantic>=2.5.0
aiofiles==23.2.1
python-dotenv==1.0.0
langchain-gigachat>=0.3.12
